refactor: adapt for and add collapsed mode in PP
