+++ packages/sn-filter-pane/src/ext/property-panel/data/data-panel/presentation/styling-definitions/styling-panel-def.js
@@ -1,51 +1,53 @@
 const stylingPanelDef = {
-  component: 'styling-panel',
-  chartType: 'filterpane',
-  translation: 'LayerStyleEditor.component.styling',
-  chartTitle: 'Object.FilterPane',
-  useBackground: true,
-  subtitle: 'LayerStyleEditor.component.styling',
-  useGeneral: true,
-  items: {
-    headerSection: {
-      component: 'panel-section',
-      translation: 'properties.Header',
-      items: {
-        headerItems: {
-          component: 'items',
-          ref: 'components',
-          key: 'listBox',
-          items: {
-            fontWrapperItem: {
-              component: 'inline-wrapper',
-              items: {
-                fontSizeItem: {
-                  component: 'integer',
-                  ref: 'header.fontSize',
-                  translation: 'properties.fontSize',
-                  width: 9,
-                  min: 5,
-                  max: 300,
-                  defaultValue(item, data, args) {
-                    const currentTheme = args.theme.current();
-                    return parseInt(
-                      currentTheme.object?.listBox?.title?.main?.fontSize ?? currentTheme.fontSize,
-                      10,
-                    );
+  stylingPanel: {
+    component: 'styling-panel',
+    chartType: 'filterpane',
+    translation: 'LayerStyleEditor.component.styling',
+    chartTitle: 'Object.FilterPane',
+    useBackground: true,
+    subtitle: 'LayerStyleEditor.component.styling',
+    useGeneral: true,
+    items: {
+      headerSection: {
+        component: 'panel-section',
+        translation: 'properties.Header',
+        items: {
+          headerItems: {
+            component: 'items',
+            ref: 'components',
+            key: 'listBox',
+            items: {
+              fontWrapperItem: {
+                component: 'inline-wrapper',
+                items: {
+                  fontSizeItem: {
+                    component: 'integer',
+                    ref: 'header.fontSize',
+                    translation: 'properties.fontSize',
+                    width: 9,
+                    min: 5,
+                    max: 300,
+                    defaultValue(item, data, args) {
+                      const currentTheme = args.theme.current();
+                      return parseInt(
+                        currentTheme.object?.listBox?.title?.main?.fontSize ?? currentTheme.fontSize,
+                        10,
+                      );
+                    },
+                    change(data) {
+                      data.header.fontSize = !data.header.fontSize
+                        ? data.header.fontSize
+                        : Math.max(5, Math.min(300, Math.floor(data.header.fontSize)));
+                    },
                   },
-                  change(data) {
-                    data.header.fontSize = !data.header.fontSize
-                      ? data.header.fontSize
-                      : Math.max(5, Math.min(300, Math.floor(data.header.fontSize)));
-                  },
-                },
-                fontColorItem: {
-                  show: true,
-                  ref: 'header.fontColor',
-                  component: 'color-picker',
-                  defaultValue(item, data, args) {
-                    const currentTheme = args.theme.current();
-                    return { color: currentTheme.object?.listBox?.title?.main?.color ?? currentTheme.color };
+                  fontColorItem: {
+                    show: true,
+                    ref: 'header.fontColor',
+                    component: 'color-picker',
+                    defaultValue(item, data, args) {
+                      const currentTheme = args.theme.current();
+                      return { color: currentTheme.object?.listBox?.title?.main?.color ?? currentTheme.color };
+                    },
                   },
                 },
               },
@@ -53,47 +55,47 @@ const stylingPanelDef = {
           },
         },
       },
-    },
-    contentSection: {
-      component: 'panel-section',
-      translation: 'properties.Content',
-      items: {
-        contentItems: {
-          component: 'items',
-          ref: 'components',
-          key: 'listBox',
-          items: {
-            contentFontWrapper: {
-              component: 'inline-wrapper',
-              items: {
-                contentFontSize: {
-                  component: 'integer',
-                  ref: 'content.fontSize',
-                  translation: 'properties.fontSize',
-                  width: 9,
-                  min: 5,
-                  max: 300,
-                  defaultValue(item, data, args) {
-                    const currentTheme = args.theme.current();
-                    return parseInt(
-                      currentTheme.object?.listBox?.content?.fontSize ?? currentTheme.fontSize,
-                      10,
-                    );
+      contentSection: {
+        component: 'panel-section',
+        translation: 'properties.Content',
+        items: {
+          contentItems: {
+            component: 'items',
+            ref: 'components',
+            key: 'listBox',
+            items: {
+              contentFontWrapper: {
+                component: 'inline-wrapper',
+                items: {
+                  contentFontSize: {
+                    component: 'integer',
+                    ref: 'content.fontSize',
+                    translation: 'properties.fontSize',
+                    width: 9,
+                    min: 5,
+                    max: 300,
+                    defaultValue(item, data, args) {
+                      const currentTheme = args.theme.current();
+                      return parseInt(
+                        currentTheme.object?.listBox?.content?.fontSize ?? currentTheme.fontSize,
+                        10,
+                      );
+                    },
+                    change(data) {
+                      data.content.fontSize = !data.content.fontSize
+                        ? data.content.fontSize
+                        : Math.max(5, Math.min(300, Math.floor(data.content.fontSize)));
+                    },
                   },
-                  change(data) {
-                    data.content.fontSize = !data.content.fontSize
-                      ? data.content.fontSize
-                      : Math.max(5, Math.min(300, Math.floor(data.content.fontSize)));
-                  },
-                },
-                contentFontColor: {
-                  show: true,
-                  ref: 'content.fontColor',
-                  type: 'object',
-                  component: 'color-picker',
-                  defaultValue(item, data, args) {
-                    const currentTheme = args.theme.current();
-                    return { color: currentTheme.object?.listBox?.content?.color ?? currentTheme.color };
+                  contentFontColor: {
+                    show: true,
+                    ref: 'content.fontColor',
+                    type: 'object',
+                    component: 'color-picker',
+                    defaultValue(item, data, args) {
+                      const currentTheme = args.theme.current();
+                      return { color: currentTheme.object?.listBox?.content?.color ?? currentTheme.color };
+                    },
                   },
                 },
               },
