+++ packages/sn-filter-pane/src/ext/property-panel/settings/index.ts
@@ -1,0 +1,3 @@
+import settings from './settings';
+
+export default settings;
